+++ Library/MapMarker.cs
@@ -1,6 +1,6 @@
 namespace Library
 {
-	public class MapMarker(string icon, string label, uint spaceFormID = 0x00000000, double x = 0, double y = 0)
+	public class MapMarker(string icon, string label, uint spaceFormID = 0x00000000, float x = 0, float y = 0)
 	{
 		public string Icon { get; } = icon;
 
@@ -8,8 +8,8 @@ namespace Library
 
 		public uint SpaceFormID { get; } = spaceFormID;
 
-		public double X { get; } = x;
+		public float X { get; } = x;
 
-		public double Y { get; } = y;
+		public float Y { get; } = y;
 	}
 }
